final-pjt-back/requirements.txt
asgiref==3.3.1
backcall==0.1.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.1
cryptography==40.0.2
decorator==4.4.0
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.1.3
django-allauth==0.54.0
django-bootstrap5==2.0.0
django-bulma==0.5.6
django-cors-headers==3.5.0
django-extensions==3.0.9
django-jsonfield==1.4.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
parso==0.4.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycparser==2.21
Pygments==2.4.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==4.3.2
tzdata==2023.3
urllib3==1.26.2
wcwidth==0.1.7
